+++ example/android/app/src/main/java/com/transistorsoft/flutter/backgroundfetch/backgroundfetchexample/Application.java
@@ -1,12 +1,20 @@
 package com.transistorsoft.flutter.backgroundfetch.backgroundfetchexample;
 
-import android.os.StrictMode;
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+
+import com.transistorsoft.flutter.backgroundfetch.HeadlessTask;
 
 import io.flutter.app.FlutterApplication;
+import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.plugin.common.MethodCall;
+import io.flutter.plugin.common.MethodChannel;
 
 public class Application  extends FlutterApplication {
     @Override
     public void onCreate() {
+
         // Strict mode.
         /*
         StrictMode.setThreadPolicy(new StrictMode.ThreadPolicy.Builder()
@@ -25,5 +33,26 @@ public class Application  extends FlutterApplication {
         */
 
         super.onCreate();
+
+        Log.d("TSBackgroundFetch", "*********************** MainApplication");
+
+        HeadlessTask.onInitialized(new HeadlessTask.OnInitializedCallback() {
+            @Override
+            public void onInitialized(FlutterEngine engine) {
+                Log.d("TSBackgroundFetch", "********* engine started: " + engine);
+                new MethodChannel(engine.getDartExecutor().getBinaryMessenger(), "channel_foo").setMethodCallHandler(new MethodChannel.MethodCallHandler() {
+                    @Override
+                    public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
+                        Log.d("TSBackgroundFetch", "**************** Application method call handler: " + call.method);
+                        result.success(true);
+                    }
+                });
+            }
+        });
+
+
+        //MethodChannel(engine.dartExecutor.binaryMessenger, CHANNEL).setMethodCallHandler { call, result ->
+        //        handleMethod(call, result, this)
+
     }
 }