+++ ios/TSBackgroundFetch.framework/Headers/TSBackgroundFetch.h
@@ -8,6 +8,7 @@
 
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
+#import <BackgroundTasks/BackgroundTasks.h>
 
 @interface TSBackgroundFetch : NSObject
 
@@ -16,17 +17,21 @@
 @property (readonly) BOOL active;
 
 + (TSBackgroundFetch *)sharedInstance;
+-(void) registerBackgroundFetchTask:(NSString*)identifier;
+-(void) registerBackgroundProcessingTask:(NSString*)identifier;
+
 -(void) performFetchWithCompletionHandler:(void (^)(UIBackgroundFetchResult))handler applicationState:(UIApplicationState)state;
 -(void) configure:(NSDictionary*)config callback:(void(^)(UIBackgroundRefreshStatus status))callback;
 -(void) configure:(NSDictionary*)config;
--(void) addListener:(NSString*)componentName callback:(void (^)(void))callback;
+-(void) addListener:(NSString*)componentName callback:(void (^)(NSString*))callback;
 -(void) removeListener:(NSString*)componentName;
 -(BOOL) hasListener:(NSString*)componentName;
--(void) start:(void(^)(UIBackgroundRefreshStatus status))callback;
--(void) start;
--(void) stop;
--(void) finish:(NSString*)tag result:(UIBackgroundFetchResult) result;
+-(void) start:(NSString*)identifier callback:(void(^)(UIBackgroundRefreshStatus status))callback;
+-(void) start:(NSString*)identifier;
+-(void) stop:(NSString*)identifier;
+-(void) finish:(NSString*)tag;
 -(void) status:(void(^)(UIBackgroundRefreshStatus status))callback;
+-(NSError*) scheduleTask:(NSString*)taskId delay:(NSTimeInterval)delay callback:(void(^)(NSString* taskId))callback;
 
 @end