+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## 0.5.0 - 2020-01-03
+* [Changed] Migrate iOS deprecated "background-fetch" API to new [BGTaskScheduler](https://developer.apple.com/documentation/backgroundtasks/bgtaskscheduler?language=objc).  See new required steps in iOS Setup.
+* [Added] Added new `BackgroundFetch.scheduleTask` method for scheduling custom "onehot" and periodic tasks in addition to the default fetch-task.
+```dart
+BackgroundFetch.configure(BackgroundFetchConfig(
+  minimumFetchInterval: 15,
+  stopOnTerminate: false
+), (String taskId) {  // <-- [NEW] taskId provided to Callback
+  print("[BackgroundFetch] taskId: $taskId");
+  switch(taskId) {
+    case 'foo':
+      // Handle scheduleTask 'foo'
+      break;
+    default:
+      // Handle default fetch event.
+      break;
+  }
+  BackgroundFetch.finish(taskId);  // <-- [NEW] Provided taskId to #finish method.
+});
+
+// This event will end up in Callback provided to #configure above.
+BackgroundFetch.scheduleTask(TaskConfig(
+  taskId: 'foo',  //<-- required
+  delay: 60000,
+  periodic: false  
+));
+```
+
+## Breaking Changes
+* With the introduction of ability to execute custom tasks via `#scheduleTask`, all tasks are executed in the Callback provided to `#configure`.  As a result, this Callback is now provided an argument `String taskId`.  This `taskId` must now be provided to the `#finish` method, so that the SDK knows *which* task is being `#finish`ed.
+
+```dart
+BackgroundFetch.configure(BackgroundFetchConfig(
+  minimumFetchInterval: 15,
+  stopOnTerminate: false
+), (String taskId) {  // <-- [NEW] taskId provided to Callback
+  print("[BackgroundFetch] taskId: $taskId");
+  BackgroundFetch.finish(taskId);  // <-- [NEW] Provided taskId to #finish method.
+});
+``` 
+
 ## 0.4.0 - 2019-12-17
 * [Changed] Upgrade to new Flutter Plugin API "V2".  Requires flutter sdk version 1.12.  See [Upgrading pre 1.12 Android Projects](https://github.com/flutter/flutter/wiki/Upgrading-pre-1.12-Android-projects)